+++ experimental/batch_process/batch_process.py
@@ -4,8 +4,8 @@ from typing import Optional, Union
 
 from freemocap.core_processes.process_motion_capture_videos.process_recording_folder import process_recording_folder
 from freemocap.export_data.blender_stuff.export_to_blender import export_to_blender
-from freemocap.parameter_info_models.recording_info_model import RecordingInfoModel
-from freemocap.parameter_info_models.recording_processing_parameter_models import RecordingProcessingParameterModel
+from freemocap.recording_models.recording_info_model import RecordingInfoModel
+from freemocap.recording_models.post_processing_parameter_models import PostProcessingParameterModel
 from freemocap.system.paths_and_filenames.path_getters import get_blender_file_path
 
 
@@ -49,7 +49,7 @@ def process_recording_without_gui(
     path_to_camera_calibration_toml: Optional[Union[str, Path]] = None,
     path_to_blender_executable: Optional[Union[str, Path]] = None,
 ):
-    rec = RecordingProcessingParameterModel()
+    rec = PostProcessingParameterModel()
 
     rec.recording_info_model = RecordingInfoModel(recording_folder_path=Path(recording_path))
