+++ freemocap/core_processes/detecting_things_in_2d_images/mediapipe_stuff/mediapipe_skeleton_detector.py
@@ -14,7 +14,7 @@ from freemocap.core_processes.detecting_things_in_2d_images.mediapipe_stuff.medi
 from freemocap.core_processes.detecting_things_in_2d_images.mediapipe_stuff.mediapipe_skeleton_names_and_connections import (
     mediapipe_tracked_point_names_dict,
 )
-from freemocap.parameter_info_models.recording_processing_parameter_models import MediapipeParametersModel
+from freemocap.recording_models.post_processing_parameter_models import MediapipeParametersModel
 from freemocap.system.paths_and_filenames.file_and_folder_names import MEDIAPIPE_2D_NPY_FILE_NAME, \
     ANNOTATED_VIDEOS_FOLDER_NAME, \
     MEDIAPIPE_BODY_WORLD_FILE_NAME
