+++ freemocap/gui/qt/widgets/control_panel/process_mocap_data_panel/process_motion_capture_data_panel.py
@@ -21,8 +21,8 @@ from freemocap.gui.qt.widgets.control_panel.process_mocap_data_panel.parameter_g
 from freemocap.gui.qt.workers.process_motion_capture_data_thread_worker import (
     ProcessMotionCaptureDataThreadWorker,
 )
-from freemocap.parameter_info_models.recording_processing_parameter_models import (
-    RecordingProcessingParameterModel,
+from freemocap.recording_models.post_processing_parameter_models import (
+    PostProcessingParameterModel,
 )
 
 logger = logging.getLogger(__name__)
@@ -33,7 +33,7 @@ class ProcessMotionCaptureDataPanel(QWidget):
 
     def __init__(
             self,
-            recording_processing_parameters: RecordingProcessingParameterModel,
+            recording_processing_parameters: PostProcessingParameterModel,
             get_active_recording_info: Callable,
             kill_thread_event: threading.Event,
             parent=None,
@@ -94,7 +94,7 @@ class ProcessMotionCaptureDataPanel(QWidget):
     def _add_parameters_to_parameter_tree_widget(
             self,
             parameter_tree_widget: ParameterTree,
-            session_processing_parameter_model: RecordingProcessingParameterModel,
+            session_processing_parameter_model: PostProcessingParameterModel,
     ):
         parameter_group = self._convert_session_processing_parameter_model_to_parameter_group(
             session_processing_parameter_model
@@ -105,7 +105,7 @@ class ProcessMotionCaptureDataPanel(QWidget):
 
     def _convert_session_processing_parameter_model_to_parameter_group(
             self,
-            session_processing_parameter_model: RecordingProcessingParameterModel,
+            session_processing_parameter_model: PostProcessingParameterModel,
     ):
 
         return Parameter.create(
@@ -149,7 +149,7 @@ class ProcessMotionCaptureDataPanel(QWidget):
 
     def _create_session_parameter_model(
             self,
-    ) -> RecordingProcessingParameterModel:
+    ) -> PostProcessingParameterModel:
         recording_processing_parameter_model = extract_parameter_model_from_parameter_tree(parameter_object=self._parameter_group)
         recording_processing_parameter_model.recording_info_model = self._get_active_recording_info()
