+++ freemocap/gui/qt/workers/process_motion_capture_data_thread_worker.py
@@ -6,7 +6,7 @@ from PyQt6.QtCore import pyqtSignal, QThread
 from freemocap.core_processes.process_motion_capture_videos.process_recording_folder import (
     process_recording_folder,
 )
-from freemocap.parameter_info_models.recording_processing_parameter_models import RecordingProcessingParameterModel
+from freemocap.recording_models.post_processing_parameter_models import PostProcessingParameterModel
 
 logger = logging.getLogger(__name__)
 
@@ -15,7 +15,7 @@ class ProcessMotionCaptureDataThreadWorker(QThread):
     finished = pyqtSignal()
     in_progress = pyqtSignal(str)
 
-    def __init__(self, session_processing_parameters: RecordingProcessingParameterModel,
+    def __init__(self, session_processing_parameters: PostProcessingParameterModel,
                  kill_event: multiprocessing.Event, parent=None):
         super().__init__()
         self._session_processing_parameters = session_processing_parameters
